# core/openapi.yaml
openapi: "3.1.0"
info:
  title: Robin API
  version: "0.3.0"
  description: REST API for Robin.OS — a personal knowledge management system.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /entries/{id}/fragments:
    get:
      operationId: listEntryFragments
      summary: Get all fragments derived from an entry
      tags: [Entries]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Fragments for this entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fragmentListResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"

  # ── Fragments ──────────────────────────────────────────────────────
  /fragments:
    get:
      operationId: listFragments
      summary: List fragments
      tags: [Fragments]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: List of fragments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fragmentListResponseSchema"
    post:
      operationId: createFragment
      summary: Create a new fragment
      tags: [Fragments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createFragmentBodySchema"
      responses:
        "201":
          description: Created fragment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fragmentWithContentResponseSchema"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /fragments/{id}:
    get:
      operationId: getFragment
      summary: Get a fragment by ID (includes content)
      tags: [Fragments]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Fragment with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fragmentWithContentResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
    put:
      operationId: updateFragment
      summary: Update a fragment
      tags: [Fragments]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateFragmentBodySchema"
      responses:
        "200":
          description: Updated fragment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fragmentResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"

  # ── Threads ────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by better-auth after sign-in
    hmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature of request body
    mcpToken:
      type: apiKey
      in: query
      name: token
      description: JWT token for MCP access

  # ── Shared Parameters ──────────────────────────────────────────────
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Resource ID or lookup key
      schema:
        type: string
...
  schemas:
...
    errorResponseSchema:
      type: object
      properties:
        error:
          type: string
        fields: {}
      required: [error]
      additionalProperties: false
...
    createFragmentBodySchema:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        entryId:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
          default: []
      required: [title, entryId]
      additionalProperties: false
    updateFragmentBodySchema:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
      additionalProperties: false
...
    fragmentResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        title:
          type: string
        type:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        entryId:
          type: string
          nullable: true
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, lookupKey, userId, slug, title, type, tags, entryId, state, repoPath, createdAt, updatedAt]
      additionalProperties: false
    fragmentWithContentResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        title:
          type: string
        type:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        entryId:
          type: string
          nullable: true
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        content:
          type: string
      required: [id, lookupKey, userId, slug, title, type, tags, entryId, state, repoPath, createdAt, updatedAt, content]
      additionalProperties: false
    fragmentListResponseSchema:
      type: object
      properties:
        fragments:
          type: array
          items:
            $ref: "#/components/schemas/fragmentResponseSchema"
      required: [fragments]
      additionalProperties: false

    # ── Threads ────────────────────────────────────────────────────
...